removing brax/common/ and moving all files to brax/
